requirement.txt
pandas
numpy
pymongo
scikit-learn
dotenv
pyyaml
dill

#-e.